+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 v1.2.0 (2013/09/27)
 
 * Upgrade jQuery to 1.10.2 and 2.0.3
-* Add new device (LG Nexus 4)
+* Add new devices (LG Nexus 4, Samsung Y, Samsung S)
 * Extend the width of preview 15px (scroll bar width)
 * Remove screenshot feature
+* [#8] Use real device width
 
 v1.1.1 (2013/08/22)